LayoutTests/pointer-lock/pointer-lock-api-expected.txt

PASS navigator.webkitPointer is defined.
PASS typeof navigator.webkitPointer.isLocked is 'boolean'
PASS navigator.webkitPointer.isLocked is false
PASS navigator.webkitPointer.lock() threw exception TypeError: Not enough arguments.
PASS navigator.webkitPointer.lock(document) is undefined
PASS navigator.webkitPointer.unlock() is undefined
PASS successfullyParsed is true

TEST COMPLETE
